+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 0.2.0
+* Add some keybindings
+  * `ctrl-x ctrl-s` - `core:save-note`
+  * `alt-<` - `emacs-mode:beginning-of-buffer`
+  * `alt->` - `emacs-mode:end-of-buffer`
+
 ## 0.1.0 - First Release
 * Miminal features